test: Move out in-source tests, refactoring with fixtures
