+++ src/client/index.ts
@@ -1,9 +1,1 @@
 export { Raindrop } from './client';
-
-/* c8 ignore start */
-if (import.meta.vitest) {
-	const { it } = import.meta.vitest;
-
-	it.skip('Nothing to test here');
-}
-/* c8 ignore stop */
