+++ src/client/raindrop.ts
@@ -1,7 +1,5 @@
-import raindropResponse from '^/tests/fixtures/raindrops.json';
-import axios, { AxiosRequestConfig } from 'axios';
-import MockAdapter from 'axios-mock-adapter';
-import { Configuration, Raindrop, RaindropApi as _RaindropApi } from '~/generated';
+import { AxiosRequestConfig } from 'axios';
+import { Raindrop, RaindropApi as _RaindropApi } from '~/generated';
 
 export interface GetAllRaindropsParams {
 	/** Sort option. See https://developer.raindrop.io/v1/raindrops/multiple#common-parameters */
@@ -59,90 +57,3 @@ export class RaindropApi extends _RaindropApi {
 		return items;
 	}
 }
-
-/* c8 ignore start */
-if (import.meta.vitest) {
-	const { describe, expect, it, afterEach } = import.meta.vitest;
-
-	// Mocks
-	const mockAxios = new MockAdapter(axios, { onNoMatch: 'throwException' });
-	const mockPaginatedResponse = (page: number, pageSize: number) => {
-		return {
-			...raindropResponse,
-			items: raindropResponse.items.slice(page * pageSize, (page + 1) * pageSize)
-		};
-	};
-
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	const getParams = (url: string) => Object.fromEntries(new URL(url).searchParams.entries());
-
-	const raindropApi = new RaindropApi(new Configuration(), undefined, axios);
-
-	afterEach(() => {
-		mockAxios.resetHandlers();
-		mockAxios.resetHistory();
-	});
-
-	describe(raindropApi.getAllRaindrops, () => {
-		it('fetch full pagination results', async () => {
-			mockAxios.onGet(/https:\/\/api\.raindrop\.io\/rest\/v1\/raindrops\/0/).reply((config) => {
-				const params = getParams(config.url ?? '');
-				expect(params).toEqual({
-					sort: '',
-					perpage: '1',
-					page: expect.any(String),
-					search: ''
-				});
-				const response = mockPaginatedResponse(parseInt(params.page), 1);
-				return [200, response];
-			});
-			const result = await raindropApi.getAllRaindrops(0, { pageSize: 1 });
-			expect(mockAxios.history.get.length).toBe(5);
-			expect(result).toHaveLength(5);
-			expect(result).toEqual(raindropResponse.items);
-		});
-
-		it('single page size', async () => {
-			mockAxios.onGet(/https:\/\/api\.raindrop\.io\/rest\/v1\/raindrops\/0/).reply((config) => {
-				const params = getParams(config.url ?? '');
-				expect(params).toEqual({
-					sort: '',
-					perpage: '5',
-					page: '0',
-					search: ''
-				});
-				return [200, mockPaginatedResponse(0, 5)];
-			});
-			const result = await raindropApi.getAllRaindrops(0, { pageSize: 5 });
-			expect(mockAxios.history.get.length).toBe(1);
-			expect(result).toEqual(raindropResponse.items);
-		});
-
-		it('collection not found', () => {
-			mockAxios
-				.onGet(/https:\/\/api\.raindrop\.io\/rest\/v1\/raindrops\/35947374/)
-				.reply((config) => {
-					const params = getParams(config.url ?? '');
-					expect(params).toEqual({
-						sort: '',
-						perpage: '50',
-						page: '0',
-						search: ''
-					});
-					return [
-						403,
-						{
-							result: false,
-							status: 403,
-							errorMessage: 'collection 35947374 not found',
-							access: false
-						}
-					];
-				});
-			expect(raindropApi.getAllRaindrops(35947374)).rejects.toThrowError(
-				'Request failed with status code 403'
-			);
-		});
-	});
-}
-/* c8 ignore stop */
