+++ src/index.ts
@@ -6,11 +6,3 @@ export { client, generated, utils };
 
 // Handy default
 export default client.Raindrop;
-
-/* c8 ignore start */
-if (import.meta.vitest) {
-	const { it } = import.meta.vitest;
-
-	it.skip('Nothing to test here');
-}
-/* c8 ignore stop */
