+++ src/utils/index.ts
@@ -1,11 +1,3 @@
 import * as tree from './tree';
 
 export { tree };
-
-/* c8 ignore start */
-if (import.meta.vitest) {
-	const { it } = import.meta.vitest;
-
-	it.skip('Nothing to test here');
-}
-/* c8 ignore stop */
